+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - `Synchronizer.getNearestRewindHeight` (its function is now handled internally
   by `Synchronizer.rewindToNearestHeight`).
+- `Synchronizer.quickRewind` and `CompactBlockProcessor.quickRewind` have been removed as they triggered the block 
+  rewind action at an invalid height. Use `Synchronizer.rewindToNearestHeight` instead.
 
 ## [2.2.4] - 2024-09-16